1.3

+++ library/src/main/java/com/common/WidgetExtensions.kt
@@ -14,12 +14,12 @@ import com.baby.viewtools.ConfirmDialog
 fun MotionEvent.isTouchIn(rect: Rect, offsetX: Int = 0, offsetY: Int = 0): Boolean {
     val x = x + offsetX
     val y = y + offsetY
-    return x in rect.left..rect.right && y in rect.top..rect.bottom
+    return x.toInt() in rect.left..rect.right && y.toInt() in rect.top..rect.bottom
 }
 
 fun MotionEvent.isTouchInX(rect: Rect, offsetX: Int = 0, spreadTouchRange: Int = 0): Boolean {
     val x = x + offsetX
-    return x in (rect.left - spreadTouchRange)..(rect.right + spreadTouchRange)
+    return x.toInt() in (rect.left - spreadTouchRange)..(rect.right + spreadTouchRange)
 }
 
 fun Drawable.toBitmap(
